requirements.txt
streamlit
scikit-learn
pandas
openai
python-dotenv
